have insufficient printing working
