+++ gamewinner/printers/ifile_printer.py
@@ -1,4 +1,5 @@
 import os
+from abc import abstractmethod
 from pathlib import Path
 from typing import Any, TextIO
 
@@ -8,12 +9,13 @@ from gamewinner.printers.iprinter import IPrinter
 
 class IFilePrinter(IPrinter):
     extension: str = NotImplemented
+    name: str = NotImplemented
 
     @classmethod
     def print(cls, bracket: Bracket, *args: Any, **kwargs: Any) -> None:
         dest_dir = Path(__file__).parent.parent.parent.joinpath("generated")
         filename = dest_dir.joinpath(
-            f"{bracket.strategy.name}-{bracket.year.year}-{cls.name}.{cls.extension}"
+            f"{bracket.strategy}-{bracket.year}-{cls.name}.{cls.extension}"
         )
         if filename.exists():
             os.remove(filename)
@@ -21,5 +23,6 @@ class IFilePrinter(IPrinter):
             cls._print(f, bracket)
 
     @classmethod
+    @abstractmethod
     def _print(cls, fd: TextIO, bracket: Bracket) -> None:
-        raise NotImplementedError
+        return NotImplemented
